backend/requirements.txt
Flask==3.1.2
flask-cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
pandas>=2.0.0
scikit-learn>=1.0.0
joblib==1.5.1
requests==2.32.3
RapidFuzz==3.13.0
python-dotenv==1.1.1
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
Werkzeug>=3.0.0
